simplify demo and data synchronisation
