--- sessions/react-data-fetching/demo-start/components/Joke/index.js
@@ -1,46 +1,0 @@
-import { useEffect, useState } from "react";
-
-export default function Joke() {
-  const [joke, setJoke] = useState();
-  const [id, setId] = useState(0);
-
-  useEffect(() => {
-    async function startFetching() {
-      const response = await fetch(
-        `https://example-apis.vercel.app/api/bad-jokes/${id}`
-      );
-      const joke = await response.json();
-
-      setJoke(joke);
-    }
-
-    startFetching();
-  }, [id]);
-
-  function handlePrevJoke() {
-    setId(joke.prevId);
-  }
-
-  function handleNextJoke() {
-    setId(joke.nextId);
-  }
-
-  if (!joke) {
-    return <h1>Loading...</h1>;
-  }
-
-  return (
-    <>
-      <small>ID: {id}</small>
-      <h1>{joke.joke}</h1>
-      <div>
-        <button type="button" onClick={handlePrevJoke}>
-          ← Prev Joke
-        </button>
-        <button type="button" onClick={handleNextJoke}>
-          Next Joke →
-        </button>
-      </div>
-    </>
-  );
-}
