+++ sessions/react-data-fetching/demo-start/pages/index.js
@@ -1,5 +1,49 @@
-import Joke from "../components/Joke";
+import { useEffect, useState } from "react";
 
 export default function HomePage() {
-  return <Joke />;
+  const [joke, setJoke] = useState();
+  const [id, setId] = useState(0);
+
+  useEffect(() => {
+    async function startFetching() {
+      const response = await fetch(
+        `https://example-apis.vercel.app/api/bad-jokes/${id}`
+      );
+      const joke = await response.json();
+
+      setJoke(joke);
+    }
+
+    startFetching();
+  }, [id]);
+
+  function handlePrevJoke() {
+    setId(joke.prevId);
+  }
+
+  function handleNextJoke() {
+    setId(joke.nextId);
+  }
+
+  if (!joke) {
+    return <h1>Loading...</h1>;
+  }
+
+  return (
+    <>
+      <small>ID: {id}</small>
+      <h1>{joke.joke}</h1>
+      <button type="button" onClick={() => handleToggleFunnyJoke(id)}>
+        {isFunny ? "😂 is Funny" : "😐 is not Funny"}
+      </button>
+      <div>
+        <button type="button" onClick={handlePrevJoke}>
+          ← Prev Joke
+        </button>
+        <button type="button" onClick={handleNextJoke}>
+          Next Joke →
+        </button>
+      </div>
+    </>
+  );
 }
